Initial release of Recipe app
Overhaul of the app brings it up to working status with recipe editing,
listing out recipes on the home page, and displaying recipe details.

In one part I pared down the concept for RecipeDirection model,
instead just making a single HTMLField on the Recipe to contain the
directions. I figure we can just write in what we want there and
not have to screw around with the implementation too much.
Trying it the other way was costing me a whole lot of time for
really no benefit at all.

Currently recipes are editable from the admin, with ability to add
ingredients in an inline model, select ingredients by type from a
dropdown menu, choose units of measure, and even include a preparation
step.

The detail template will output a basic header, the description (if any,
though I forced that not to be printable already), the ingredients list,
directions, and footnotes (if any).

A basic stylesheet has been included which targets this page only,
`recipe_detail.css`. Some crude adjustments to the ingredients list are
included, but mainly it has stubs for all the classes that are
made available in that template's markup, hopefully so Molly can come
in later and style the place up a bit.

Finally, adds the links for a recipe detail into the recipe list page,
as well as activates the link to recipes list from the home page.

Future updates can add all the bells and whistles of searching and
whatnot, but for now we'll stick to the basics.