CSS adjustments
Created a base CSS sheet for recipes stuff,
which is expected to be imported to other stylesheets
nearby. This provides a useful hook for styles to apply
to every page in the recipes app, without needing to
ensure that stylesheet gets included in the head.

Also made some small adjustments to recipe_detail.html template.
Think this one's finalized until we get Molly in here
to look at the styling.