Database backup and restore scripts using AWS S3
Backup script dumps database to file, then uses awscli to copy that file
up to an S3 bucket for safe keeping.

The restore script is able to read the latest backup off the S3 bucket,
copy it to the system, and run pg_restore with it.

Finally, ls_latest will simply read the latest file off the bucket.

See config.sh.template for template of a config script
necessary for each of these other scripts to run.
It should be copied as "config.sh" and changed to fit
production needs.

The real config file is ignored in gitignore, as are
all dump files the backup script would generate.