Settings split, logging added
Separated settings into base, dev, and prod, switched on the
DEBUG flag..

For now, the only things dev and prod change are logging handlers.
Still, a framework exists here to set base settings, override them in
either dev or prod, and further override those in an optional local.py.

The caveat is that LOGGING is defined in full within __init__.py.
Parts of the config can be defined in other modules,
such as LOGGING_FORMATTERS and LOGGING_HANDLERS.

In truth, this type of setup is not a new concept,
and I'm sure someone out there has made a more robust version.
This works for us right now, though.