Reduce some of the DB calls for related objects.
I may have only reduced the calls by 1 or 2 here, compared to 6 others
that seem to be coming from django-admin-sortable. Need to dive into
that later, but for now I know I've done a bit here to ease that.

Plus, the admin is not going to be the real UX focus of the project:
eventually, we'll add some cool UI layer stuff, probably React
with DRF or Graphene on the backend, that will be the real focus
of performance scrutiny.