Add a deployment script
Doesn't fit well as a management command, so leaving here.

gitignore similar to the config.sh pattern for db backups,
with a template included here to copy and adjust on the server.

Deployment should be relatively straightforward, checking out code,
running collectstatic and migrate,
and restarting services.

May need to add a flag to do pip installs
if there are changes in the requirements,
or just run it blind. Not sure how I want to handle that yet.