Database-level plural name for ingredient types
In case the plural of something is not easy, like "leaf" is "leaves"
and not "leafs", store an optional plural name for that type in the
model, then use that in the template output, instead.

Briefly considered turning that to a field to output the plural name,
but I think that would have to apply to the RecipeIngredient and
not the type.

In the end, that would have been too complex. This is sufficient for
our needs.