Recipe search functionality
Basic full-text searching using postgres contrib packages in Django.
Adds control to search the list of recipes into the frontend,
and the backend view will filter those results down.

May not be perfect for searching contents of directions and such, and
my understanding of search rank utterly sucks,
but it should get the job done 90% of the time.

Also, customizes the title section for the list a bit, moving away from
the title card widget. This is so the search feature
could be included in that title card more cleanly.