Add recipe ingredient sections
There are problems with it, as noted in comments in the
ingredients.html template. Still, managed properly, it works.

Moving forward, we'll just get out of the Admin side of things
and build more of a frontend with React to take care of the details
and provide a better editor experience.