Adjust tinymce config to include code editor
Plugin was activated, but not present by default.

Also, django-tinymce docs state the wrong default for this new setting.
I discovered the real default used by going into its source,
where it does a getattr on settings for this particular setting,
then actually defines it as "DEFAULT_CONFIG".
Should go back and note that to the author of that package.