Docker transformation (#18)
* Initial hunnydo app files.

* Massive changes on this one

- Removing invoices and hunnydo apps entirely: focusing on recipes site
- Removed django-environ; no longer needed; focusing on plain os.environ
- Changed psycopg2 to psycopg2-binary. Easier time running in containers (spoiler alert)
- flatpages removed
- IngredientType field plural_name changed to semi-private, now using
  property by the old name to calculate same.
- Last but not least, Dockerized services!

* Packages updated

* App name corrections in templates

* Docker compose override style

Use an nginx container on dev automatically with docker-compose up

In production, specify `-f docker-compose.yml` to ignore the override,
allowing us to use host nginx, instead.

* Renamed project directory

In keeping with change of the site from hometools to just recipes,
changed the project directory to recipesite

Far as I can tell, all other changes made here reflect that update.

* Updates to nginx image README

Better description for why and how it is used.

* Append compose file version

Services not recognized in prod.
Probably using an older format.

* Docker explicitly should build directories for logs, media, and static

Otherwise, doesn't seem to work properly on production.

* Update poetry lock to match pyproject dependencies

* Ensure mkdir commands work without error on build

* Make volumes for logs and media files

May have been what I was missing before

* Set some extra config for gunicorn

Also try binding on ipv6, see if that can help

* Change bind back to localhost

* Set container name for nginx test build

* Run the container for recipesite as current user

Do this instead of the default, which runs as root
and complicates the heck out of some stuff.

* Roll back that last change

* Missing backupcount for log files

* Set forward proto header as secure

Tells Django that the header sent by Cloudflare indicates it should be secure.

At least I think that's how it works. We'll see if it all gets screwed in the end.

* Change the admin endpoint

Probably dumped it at some point, not worth the trouble now.