Tailwind css overhaul (#28)
* Require login for main page

Sprucing up that login page to match (wip)

* Remove nginx docker stuff from repo

* Hello, Tailwind and custom user accounts

- Login by email or username available
- Tailwind CSS implemented, replaces Bootstrap
- Big BIG overhaul on:
  - login view
  - base template
  - header
  - footer
- Logo added (temp?)

This is a work in progress. Still to do:

- home page (list of recipes)
- recipe details
- Search function

There's also an "About" link that is non-functional.
Used to have a flatpages view for that.
Maybe turn that back on?

* Recipe search updated

Status of search still a work-in-progress, but getting there.

Breadcrumbs also removed entirely.
If we want them later, we can rebuild them.

* Require login for main page

Sprucing up that login page to match (wip)

* Remove nginx docker stuff from repo

* Hello, Tailwind and custom user accounts

- Login by email or username available
- Tailwind CSS implemented, replaces Bootstrap
- Big BIG overhaul on:
  - login view
  - base template
  - header
  - footer
- Logo added (temp?)

This is a work in progress. Still to do:

- home page (list of recipes)
- recipe details
- Search function

There's also an "About" link that is non-functional.
Used to have a flatpages view for that.
Maybe turn that back on?

* Recipe search updated

Status of search still a work-in-progress, but getting there.

Breadcrumbs also removed entirely.
If we want them later, we can rebuild them.

* [wip] Change flow to intro page and recipes section

Getting deeper into it with the tailwind built.

Not much success at the moment getting the prod build for tailwind
to build itself in Docker via args and env variables like I expect it.
More investigation there needed to determine how to move forward.

* Recipe search results updated

* Recipe list with pagination, updated

* AlpineJs integration: toggling nav on mobile

* [wip] Detail page component refresh

* Renamed data fixture

Easier to call in this way.

* Header fixes

- Buttons stretch and stack on small screen
- Remove unneeded links to Home and Recipes

* Recipe detail header stats fleshed out.

* Hero image template removed (unneeded)

* FInal pass, details ready

Some lingering issues that can be fixed later:

- style of prep and cook times don't match total time output
- number of servings listed, but does nothing
- did not yet check for nutrition label output
- hero images not yet functioning

Other than those issues, good to go!

* Poetry dependency updates