feat: change from Alpine to Hyperscript
Alpine proved a bit troublesome for its use of eval,
requiring special CSP rules to execute.
Hyperscript is locally served in this case, and makes more sense for my
usage.

Getting rid of django-csp generally, as well. This is not something
the app should be controlling: better to leave it at the web server
level (Nginx and such) where there are better standards, I think.