chore: dependency updates
Uvicorn standard extra adds some optionals I don't see a reason not
to include.

Setting the PATCH version for compatibility of
django-admin-sortable2
should make that a bit more stable.
Honestly I'm considering ripping that out as it is, but we'll see where
things go later on.

And added the description because that's always nice.