ci(wip): 👷 refactors on CI workflow
Removed the duplicated docker image build push, as ci.yaml does that, as
well. We're relying now on pre-commit.ci to handle those checks, so no
need for the duplication. And setting up some initial stuff for testing,
but running into trouble getting pytest-django to play nice with
djfractions (it doesn't see my monkey patches). So, no joy on actually
running any tests, but maybe we can poke the maintainer of djfractions
to get something released?