feat: recipe card list design finalized
Well, for now, of course. Cards layout is in columns, which is not
particularly fancy in terms of CSS styles. Had to ensure the overflow
of the cards worked without cutting parts of content off.

The downside is that cards flow DOWN first before breaking on columns,
but I don't think that will be much of an issue.

Next step is eventual pagination and search options.
For now, though, time for some content.