feat: stats layout finalized
Went back in time to the Django version of this site,
found the stats template I used there.
Granted I was using these base64-encoded SVGs for icons at the time,
but this change unravels those to the raw SVG code.

The stats include is added twice, with different breakpoints.
Initially I tried to make the grid flow differently with a row span and
other weird stuff, but ultimately this works out better.
The stats on larger screens tucks under the content, beside the image;
while on small screens it lays out beneath the image.